+++ src/menu-entry/menuDialog_ui_derived.cpp
@@ -5,6 +5,8 @@ DialogMenuEntry::DialogMenuEntry(wxWindow* parent): DialogMenuEntryBase(parent)
 
 }
 
+
+//Update position
 void DialogMenuEntry::OnButtonClick_UpdateOwnShipPosition( wxCommandEvent& event )
 {
   m_action = DialogAction::UpdateOwnShip;
@@ -17,12 +19,29 @@ void DialogMenuEntry::OnButtonClick_UpdateAisTargetPosition( wxCommandEvent& eve
   EndModal(wxID_OK);
 }
 
-void DialogMenuEntry::OnButtonClick_AddAisTarget(wxCommandEvent&)
+
+//Add AIS targets
+void DialogMenuEntry::OnButtonClick_AddClassA(wxCommandEvent& event)
 {
-  m_action = DialogAction::AddAisTarget;
+  m_action = DialogAction::AddClassA;
   EndModal(wxID_OK);
 }
 
+void DialogMenuEntry::OnButtonClick_AddClassB(wxCommandEvent& event)
+{
+  m_action = DialogAction::AddClassB;
+  EndModal(wxID_OK);
+}
+
+void DialogMenuEntry::OnButtonClick_AddARPA(wxCommandEvent& event)
+{
+  m_action = DialogAction::AddARPA;
+  EndModal(wxID_OK);
+}
+
+
+
+//Remove AIS targets
 void DialogMenuEntry::OnButtonClick_RemoveLastAisTargets( wxCommandEvent& event )
 {
   m_action = DialogAction::RemoveLastAisTarget;