+++ src/menu-entry/menuDialog_ui_derived.h
@@ -8,9 +8,11 @@ enum class DialogAction
   None = 0,
   UpdateOwnShip = 1,
   UpdateAisTarget = 2,
-  AddAisTarget = 3,
-  RemoveLastAisTarget = 4,
-  ClearAisTargets = 5
+  AddClassA = 3,
+  AddClassB = 4,
+  AddARPA = 5,
+  RemoveLastAisTarget = 6,
+  ClearAisTargets = 7
 };
 
 class DialogMenuEntry : public DialogMenuEntryBase
@@ -22,7 +24,9 @@ class DialogMenuEntry : public DialogMenuEntryBase
   protected:
     void OnButtonClick_UpdateOwnShipPosition( wxCommandEvent& event ) override;
     void OnButtonClick_UpdateAisTargetPosition( wxCommandEvent& event ) override;
-    void OnButtonClick_AddAisTarget( wxCommandEvent& event ) override;
+    void OnButtonClick_AddClassA(wxCommandEvent& event) override;
+    void OnButtonClick_AddClassB(wxCommandEvent& event) override;
+    void OnButtonClick_AddARPA(wxCommandEvent& event) override;
     void OnButtonClick_RemoveLastAisTargets( wxCommandEvent& event ) override;
     void OnButtonClick_ClearAisTargets( wxCommandEvent& event) override;
 };